word and background
This is the start of my project. I added in my word of focus which is perspective and I will add in my other objects/layers eventually that will go with this word. I am much more comfortable with GIMP after today and am excited to see what the final looks like.